Using the same logic for both ChannelNameFormatter
